changed XUID to bigint, wrote function that filters recent players to message

+++ database/migrations/2016_10_05_025447_createProfilesTable.php
@@ -17,7 +17,7 @@ class CreateProfilesTable extends Migration
             $table->primary('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('gamertag')->nullable();
-            $table->integer('xuid')->nullable();
+            $table->bigInteger('xuid')->nullable();
             $table->longText('api')->nullable();
             $table->timestamps();
         });
